Revert "Moved spec_runner into the corresponding task."
There seems to be a hidden dependency on GEM_HOME / GEM_PATH being unset
amongst different Rake tasks. This happened to work because SpecRunner was
initialized before any of the tasks were called.

I'm reverting this commit as it's rather late and I'd rather not start moving
code around at 23:00. I'll investigate further and see if this can be resolved
properly in the coming days.

This reverts commit 18b26ab525322b3ce1b907a48693fd9c90e53321.