Removed psych from pre-installed gems.
On Heroku, when pushing a Ruby app, libyaml is installed if psych is detected
in the Gemfile or bundling process (I'm not sure which). This means that if
psych is pre-installed, rubygems will cause it to be loaded vi YAML and libyaml
won't be available, which means we can't run to install psych and Heroku can't
see that psych is being used...

Heroku could just install libyaml always and this wouldn't be an issue.