Revert "Close handles opened with dlopen()"
This reverts commit 699b03ad39468a1ef8129c333f78d2ee0c338277.

This was causing the profiler tooling extension to be unloaded
so that running the profiler with eg:

  rbx -Xprofile -e 'p 1'

would cause an EXC_BAD_ACCESS on OS X Maverics at least.

Generally, we don't want to dlclose() a library that we loaded
and intend to use. We may address this by passing RTDL_NODELETE
to dlopen(), but that mode flag is not universally supported and
was added later. If we don't intend to unload the library, there's
no reason to dlclose() it before process exit.