Handle whitespace only strings in Float::from_cstr
In certain cases when a String with only whitespace (be it spaces, newlines or
a mix of that) would be converted to a Float this would result in a crash.

The added test is a bit odd but it's based on the output/behaviour as described
in https://github.com/rubinius/rubinius/issues/2955#issuecomment-36030279.

In particular String::byte_size() returned 0 for some reason (this might be a
bug on its own as far as I can tell), which would then trigger a crash in
Float::from_cstr().

Thanks to @brixen for helping me out with this one as I was, and this might be
an understatement, rather clueless on what was going on.

See #2955 for more info.