debian stable package support for rubinis-2.2.
File necessary to build a .deb package for Debian Stable(wheezy).

Rubinius 2.2.6+ or MRI 1.8/1.9 can be used to build the package. There were
changes in Rubinius prior to 2.2.4 to make this work.

Conflict & Replace MRI ruby packages.

gem is a separate alternatives command in rubygems1.8. This is broken. The gem
command is include in ruby1.9.1 package and in rubinius it is also included.
The gem command must switch in lock step with the ruby command otherwise
everything breaks.

I force package conflict until the MRI packages are fixed.

dpkg-buildflags is used for the default build flags. This fixes some common
problems and allows system specific configuration.

Description in control file by @brixen
"Rewrote the description as one paragraph to meet the formatting restrictions.
I focused on more direct technical language as I expect the audience for the
description to be knowledgable about Ruby, programming, or the technology
described generally."