Fixed JIT bug inlining static Fixnum hash.
When the JIT was inlining this code:

https://github.com/rails/rails/blob/4-0-stable/activerecord/lib/active_record/connection_adapters/postgresql/oid.rb#L276-290

into here:

https://github.com/rails/rails/blob/4-0-stable/activerecord/lib/active_record/connection_adapters/postgresql/database_statements.rb#L136-156

the JIT was statically emitting the hash value for the key 23, so when
the key 1700 was later passed, Hash#fetch was not finding the item in
the Hash and triggering the 'unknown OID' warning.

I'm still not certain that this optimization is sensible given that Object#hash
is commonly used with Hash lookup and this basically assumes Hash keys will be
homogeneous, an assumption often contravened by common Ruby code, I'm guessing.