Make the Gems cache directory configurable.
This directory can be set using the --gems-cache option when configuring
Rubinius. Using this option one can move the Gems cache directory out of the
local working tree if needed. This in turn means CI systems that wipe the local
working tree don't have to constantly re-download the same Gems.

Considering this in itself takes time and RubyGems doesn't always respond this
can save quite the time in a build process. The trade off is that you'll have
to, once in a while, prune this directory to prevent it from ever growing in
size.

Corresponding issue: #3011