Don't leak FORTIFY_SOURCE from LLVM.
When running a debug build llvm-config could potentially leak _FORTIFY_SOURCE
in to the list of CFLAGS. This in turn would break debug builds due to warnings
being treated as errors.

This particular case is platform dependant. For example, it occurs on Amazon
CentOS but not on my local Arch Linux installation. Linux distributions keep
surprising me, and not always in good ways sadly.