Removed ruby_version/_bug guards from Hash specs
Hash specs are now 1.8.7-specific. In the case of specs previously
wrapped by ruby_bug guards, we choose to preserve those specs -- which
implies that in rbx-1.3, we choose to implement the correct, patched
behavior, and not duplicate the bugs in 1.8.7. Our hope is that should
an app be migrated to rbx-1.3 that depends on 1.8.7's bugs, developers
can use rbx's toolset to identify where those dependencies are in their
code, and remove them.

This also means that when this set of core/hash specs is run against
MRI 1.8 (bin/mspec -B spec/ruby/ruby.1.8.mspec core/hash), there will
be failures. These failures remain documented in the
specs themselves as comments pointing to the redmine ticket and the
version/patch at which the bugs were fixed in MRI.