debian build update for rubinius-2.3
This changes all relevant references to the previous debian
package 'rubinius-2.2' to 'rubinius-2.3'.

The file create/rm are a product of the debian package name, which
includes the major version number, so new files had to be created
and the older ones deleted.

todo: Integrate these version number updates and package switch at
major version changes with the main rubinius release tools.