Use "status" ivar in Process::Status.
MRI stores the status code in an instance variable called "@status" instead of
"@exitstatus". Due to the lack of a proper API to set this variable people
instead have to resort to setting it via rb_ivar_set() in C. To ensure
compatibility with MRI we sadly also have to name this variable in the same way.

Due to the variable being renamed the method "exitstatus" now has to be manually
defined, instead of just being an attribute reader.

See #3268 for more information.