Removed libyaml (again).
Since late 2012, Psych vendors libyaml but only uses it if the system
libraries are missing. This causes a problem bootstrapping Rubinius
because the Ruby build dependency most likely needs libyaml installed
to run the Ruby to build Rubinius. We can work around this by using
the previous Rubinius build for platforms (like Heroku) where binaries
need to be built to run on systems without libyaml installed.