Fixed forking from multiple threads. Fixes #3302.
Only one thread is allowed to fork() at a time in order to correctly manage
system threads and other resources when calling fork(). The threads that are
waiting their turn to fork must not prevent the thread that is forking from
pausing the world so that a deterministic state can be set up for the child
process.

All this stop-the-world, GC [in]dependence, and locking needs to go DIAF.

Baby steps.