Removed compiled file version check.
The version check was necessary when Rubinius used language modes because
the signature would only uniquely identify the entire build toolchain and
virtual machine, but not the specific language mode used. This would create
an issue if running the same Ruby file in different language modes because
the bytecode differs.