Added spec for Kernel#inspect.
In MRI, Kernel#inspect checks if no instance variables are defined and
if that's true, side-calls to the C function implementing Kernel#to_s.

https://github.com/ruby/ruby/blob/e28c3d5df4f5abc83e0d2de34e7ebf675c96a307/object.c#L564

Consequently, when Kernel#to_s is overridden, that Ruby method is never
called.

In Rubinius, we used the same logic, but actually called the Ruby method
Kernel#to_s. If the method was overridden, the new method would be called.