Added RUBY_ENGINE_VERSION.
Similar to RUBY_ENGINE that Rubinius added many years ago, RUBY_ENGINE_VERSION
is the version of the Ruby engine. This tracks the feature that Bundler
provides for specifying the Ruby version, Ruby engine, and Ruby engine version
in a Gemfile. Unfortunately, Bundler chose ':engine' and ':engine_version' as
the key names.