Fix enum size/counting for each_slice & friends.
As it turns out we can't always use Enumerable#count as a fallback as this
doesn't work for either very large enumerators or lazy ones that go on to
infinity (e.g. 0..Float::INFINITY).

To work around this we _only_ use Enumerable#count as a fallback for
Enumerable#each_slice and _only_ if the enumerator is not a Enumerator::Lazy
instance. The resulting code is a bit messy but I can't think of any better
place to put it.