Updates and implements Enumerator#size calculation on Enumerable methods that return an Enumerator instance.
Moves Enumerable#each_cons and Enumerable#each_slice argument validation to be done before checking if there's no block given.
It creates the EnumerableHelper module to implement the cycle_size 'function' calculation. It avoids 'polluting' the Enumerable module and also allows the cycle_size function to be used on Array.