Changed default log location and permissions.
Typical programs that log to eg /var/log are daemons or services. There is typically
a single process or process group on a node. In contrast, Rubinius may be run by
multiple users on a single node. Segregating the log file by user enables better
control of the log file permissions.