Added unconditional logger::write.
Many log messages may be associated with a particular log level, but some
log output (eg the command line of the process) are not necessarily associated
with a log level as that is typically specified. In other words, some info
should always be written to the log. It's not a matter of program 'condition'
(eg warning, debug, error). The content *is* informational, but the 'info'
level typically indicates something between optional and important, inclusive,
and may be turned off.

Because of syslog limitations, the write() output is tagged LOG_INFO.