Reworked Console connection mechanism.
One major problem with the previous mechanism was that there's no way
for a process to clean up from a SIGKILL, so there's no way to not litter
some directory with the console files that the process would pre-emptively
create.

The new process timeline looks like this:

1. When a Rubinius process starts, it opens (creating if necessary, and
truncating) the system.console.path file. It watches that file for
modifications. If the process is killed, that file will persist, but there
is only one such file for N many processes. Opening the file advertises that
the process is available for Console connections. The file descriptor has
the close-on-exec flag set.

2. When a change is detected on the system.console.path file, the process
stats two files named ${system.console.path}-$pid-{request, response}. If
these files exist and are regular files, a connection is initiated.

3. Threads are created for the request, response, and worker functions.
While the connection is maintained by the client, new connection attempts
are ignored.

4. When the connection is terminated by the client, a new connection may
be initiated as described in 2.

When the process with a connected client forks, the child process closes
the Console request and response files.

When a process without a connected client forks, the listener is recreated
in the child.