Fix Method#==
A Method defined via def has an executable attribute of class
Rubinius::CompiledCode. On the other site a method defined via
define_method has an executable attribute of class
Rubinius::BlockEnvironment::AsMethod.

Comparing method equality includes the comparison of executable
equality.

For comparing an executable of type Rubinius::BlockEnvironment::AsMethod
against an executable of Rubinius::CompiledCode results in an error because it tries
to access the block_env attribute on an Rubinius::CompiledCode object,
which doesn't exist.

This commit adds an early return like in Rubinius::CompiledCode#==