Fixed broken SignalException spec.
This spec has two main problems:

1. It contains a race between executing Process.kill, which will signal the
process and processing that signal will raise a SignalException, and rescuing
a SignalException. There's nothing that guarantees that the signal is even
delivered by the OS before the thread of execution moves on.
2. It assumes that the signal will be processed in this thread, thereby resulting
in the exception being raised on this thread and it being possible to rescue the
exception. This assumption would fail if this code were running in a thread other
than the one processing signals.

Signals are delivered and handled asynchronously. The signal trapping mechanism
is the proper way to handle signals. The thread that the signal is processed on
is undefined in Ruby. Rubinius now processes signals on a dedicated thread that
is invisible to the other threads (ie, no user code will ever run on that thread
other than signal handler blocks).