Reworked halting the system.
This is the first step of a much refactor to introduce a proper System object
instead of hanging all the process boot and halt logic off the Environment
object. A lot of boot and halt code can then be cleaned up (eg creating and
initializing the object space, loading the kernel code, etc).

The main process thread (the one the OS creates when the process is first
created) now boots the system and spawns a thread to start running Ruby code.
The main thread goes on to process signals and halting the process. This is
the only thread that processes halt code.

When the process forks, a new thread is created to process signals and
halting. In the future, fork() and exec() will be disallowed on all internal
threads and only allowed on threads running Ruby code.