Delay tracking VM objects until Thread is running.
In the case of `Thread.new`, the OS thread will never run because a
ThreadError exception is raised when no block is passed. If we track the VM
object that would ultimately contain the reference to the OS thread, we either
need a way to remove the VM object when eg `Thread.new` raises an exception or
we will leak these objects. Instead of tracking and then untracking the VM
object, we create the object untracked and track it if the OS thread starts
executing.