Moved vm/ to machine/.
The abbreviation 'vm', usually meaning 'virtual machine', is badly overloaded
and even inconsistent now in the code base (eg with class VM, etc). So, renaming
the vm/ directory to machine/ and consistently using Rubinius [Mm]achine for
both the high-level concept of the thing that encompasses the managed runtime
(ie bytecode execution, memory management, system interface, JIT infrastructure and
native code generation) and the code that implements it.