Protect self and CompiledCode from GC before handling arguments
When arguments are keywords, `GenericArguments::call` will call
`Rubinius::Runtime.keyword_object?`, which results in allocating some
objects. When the young zone is full, garbage collection is run, which
could lead to forwarding the self and/or the CompiledCode objects
without updating their references, resulting in a faulty behavior or
crashes like in #3540.

Fixes #3540