Better integrate logger facility.
Basically, almost everything that is in util/ should not be. These major
components need to be well-integrated with the rest of Rubinius. As in the
previous case of Immix and the present case of the logger here, that means
things like taking STATE, VM, etc as parameters and interfacing with things
like process locks and process phases around fork/exec, etc.