Avoid leaking orphaned VM instances.
When starting a Thread, there is a point where resources are created but an
exception could be raised that would result in the resources not getting
reclaimed. At the moment, this is the VM instance. This will be refactored
soon, but the same caution will always apply to creating Thread instances.