A bit more CallFrame cleanup.
We don't have a constructor for CallFrame so it's not always clear which
fields are initialized. In this case, VM::call_frame_ is NULL when the VM
instance is created, so calling VM::push_call_frame sets previous to NULL when
pushing the new frame. This is fine, but brittle.

Ultimately, CallFrame needs to be split into separate frame types for managed
code, FFI, JIT, and native methods (ie C-API), and constructors should be
defined for every type.