Fixed setting internal thread size.
On FreeBSD, setting the stack to 4096 bytes for threads like the Console
Listener was not sufficient. This resulted in a SEGV while booting Rubinius,
which happens before the signal handler has been set up, so nothing is shown
in the logs. This also causes Rubinius to abort while trying to build
Melbourne, leaving Rubinius in a very difficult to debug configuration.