Improved logger locking.
The logger is a low-level mechanism that should not depend on sophisticated
Rubinius code, but still needs synchronization to produce intelligible results
when any thread at any time may log some activity.

We switch to a custom spinlock mutex implemented with C++11 features
(unfortunate that C++11 didn't think it was essential to provide such a mutex
but it's trivial to create).

Since the process may log something at any time from any thread, running
managed or unmanaged, even while fork'ing, we always reset the state of the
lock after fork().