Cleaned up passing STATE to logger.
The logging facility is intentially highly decoupled from Rubinius so that it
can be used through the entire process lifetime without depending on
structures that need to be constructed and destructed. The STATE dependency
was added to try to synchronize logging across fork() calls, but that wasn't a
good architecture choice.