Improved HAMT Hash.
The base "table" has been replaced by a regular Trie. Rubinius has deprecated
support for 32-bit and that support is being removed, so the branching factor
is currently 62 (actually 60 = 6 * 10), but that will be extended to 64 when
1. instructions are added for computing with non-mananged objects (ie not
being limited to Fixnum); and 2. managed objects with non-managed fields are
able to be processed by managed code instructions.

A linear scan list is still used for hash values that collide along (6 * 10 =
60 bits), but this could be enhanced to be a binary/red-black/etc tree
instead.