Renamed ConstantScope to LexicalScope.
Originally, this was called StaticScope, which was pretty accurate. It was
renamed ConstantScope to make it more obvious that it was intimately related
to 'constant' lookup in Ruby.

However, constants are odd things in Ruby: their values are neither constant
nor is their lookup restricted to one scope.  Instead, their values may be
reassigned and their lookup starts with the lexical scope but continues with
the inheritance chain.

With the additional of functions in Rubinius bound to lexical scope, it is
reasonable to refocus this object on its actual role: to represent a lexical
scope.