Split instructions into files.
With the expansion of the instruction set and different types of instructions
with expanded attributes, using a special syntax becomes much more complex.

Instead of managing special code to parse this special syntax, we convert to
standard JSON for attributes and documentation, and explicit C++ source files
and C #include statements for code.